requirements.txt
fastapi
flask
uvicorn[standard]
pydantic
openai
python-dotenv 